mountup/backend/requirements.txt
asgiref==3.11.1
Django==4.2.29
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
sqlparse==0.5.5
typing_extensions==4.15.0
